Moved callback structures for spin_line_console into its own class. Easier to read this way.
