--- solent/lc/cs.py
@@ -1,34 +1,0 @@
-#
-# callback structures for module lc
-#
-# // license
-# Copyright 2016, Free Software Foundation.
-#
-# This file is part of Solent.
-#
-# Solent is free software: you can redistribute it and/or modify it under the
-# terms of the GNU Lesser General Public License as published by the Free
-# Software Foundation, either version 3 of the License, or (at your option)
-# any later version.
-#
-# Solent is distributed in the hope that it will be useful, but WITHOUT ANY
-# WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
-# FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
-# details.
-#
-# You should have received a copy of the GNU General Public License along with
-# Solent. If not, see <http://www.gnu.org/licenses/>.
-
-class CsLcConnect:
-    def __init__(self):
-        self.addr = None
-        self.port = None
-
-class CsLcCondrop:
-    def __init__(self):
-        self.msg = None
-
-class CsLcCommand:
-    def __init__(self):
-        self.tokens = None
-
